% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{pdf_document_format}
\alias{pdf_document_format}
\title{Helper function to create a custom format derived from pdf_document}
\usage{
pdf_document_format(
  format,
  template = find_resource(format, "template.tex"),
  ...
)
}
\description{
Helper function to create a custom format derived from pdf_document
}
