% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{pkg_file}
\alias{pkg_file}
\title{Helper function to package template 1}
\usage{
pkg_file(...)
}
\description{
#' @title R Markdown output formats Template for thesis at DepartmentBiological Program, Univ. Tsukuba
#' @importFrom rmarkdown render
#' @importFrom rmarkdown pdf_document
#' @importFrom jpaRmd jpa_cite
#' @param Rmd_file file name of R Markdown file
#' @param Bib_file file name of Bib file
#' @export
tsukuba_thesis  <- function(Rmd_file, Bib_file) {
jpa_cite(Rmd_file, Bib_file)
tmp_rmd <- paste0("tmp_", Rmd_file)
template_tex_file <- system.file("rmarkdown/templates/thesis_tsukubabp/resources/template.tex",
package = 'TsukubaRmdBP')
format_pdf <- pdf_document(
latex_engine = "xelatex",
template = template_tex_file,
keep_tex = TRUE,
toc = TRUE,
toc_depth = 3,
highlight = "tango"
)
format_pdf$inherits <- "pdf_document"
output_file <- strsplit(Rmd_file, ".Rmd")[\link{1}]
render(tmp_rmd, format_pdf, output_file)
}
}
