% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{tsukuba_format}
\alias{tsukuba_format}
\title{R Markdown output formats Template for thesis at DepartmentBiological Program, Univ. Tsukuba}
\usage{
tsukuba_format(..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris"))
}
\description{
R Markdown output formats Template for thesis at DepartmentBiological Program, Univ. Tsukuba
}
