% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{tsukuba_thesis}
\alias{tsukuba_thesis}
\title{R Markdown output formats Template for thesis at DepartmentBiological Program, Univ. Tsukuba}
\usage{
tsukuba_thesis(Rmd_file, Bib_file)
}
\arguments{
\item{Rmd_file}{file name of R Markdown file}

\item{Bib_file}{file name of Bib file}
}
\description{
R Markdown output formats Template for thesis at DepartmentBiological Program, Univ. Tsukuba
}
